Moved all the covariance classes to edu.pitt.dbmi.ccd.math.stat.correlation package.
