+++ tetrad-lib/src/main/java/edu/cmu/tetrad/stat/correlation/CovarianceMatrix.java
@@ -16,32 +16,32 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA 02110-1301  USA
  */
-package edu.cmu.tetrad.correlation;
+package edu.cmu.tetrad.stat.correlation;
 
 /**
  *
- * Jan 27, 2016 5:35:01 PM
+ * Jan 25, 2016 2:13:26 PM
  *
  * @author Kevin V. Bui (kvb2@pitt.edu)
  */
-public class RealCovarianceMatrix implements RealCovariance {
+public class CovarianceMatrix implements Covariance {
 
-    private final double[][] data;
+    private final float[][] data;
 
     private final int numOfRows;
 
     private final int numOfCols;
 
-    public RealCovarianceMatrix(double[][] data) {
+    public CovarianceMatrix(float[][] data) {
         this.data = data;
         this.numOfRows = data.length;
         this.numOfCols = data[0].length;
     }
 
-    private double[] computeMeans() {
-        double[] mean = new double[numOfCols];
+    private float[] computeMeans() {
+        float[] mean = new float[numOfCols];
         for (int col = 0; col < numOfCols; col++) {
-            double sum = 0;
+            float sum = 0;
             for (int row = 0; row < numOfRows; row++) {
                 sum += data[row][col];
             }
@@ -52,51 +52,51 @@ public class RealCovarianceMatrix implements RealCovariance {
     }
 
     @Override
-    public double[] computeLowerTriangle(boolean biasCorrected) {
-        double[] covarianceMatrix = new double[(numOfCols * (numOfCols + 1)) / 2];
+    public float[] computeLowerTriangle(boolean biasCorrected) {
+        float[] covarianceMatrix = new float[(numOfCols * (numOfCols + 1)) / 2];
 
-        double[] mean = computeMeans();
+        float[] mean = computeMeans();
 
         int index = 0;
         for (int col = 0; col < numOfCols; col++) {
             for (int col2 = 0; col2 < col; col2++) {
-                double variance = 0;
+                float variance = 0;
                 for (int row = 0; row < numOfRows; row++) {
                     variance += ((data[row][col] - mean[col]) * (data[row][col2] - mean[col2]) - variance) / (row + 1);
                 }
-                covarianceMatrix[index++] = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+                covarianceMatrix[index++] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
             }
-            double variance = 0;
+            float variance = 0;
             for (int row = 0; row < numOfRows; row++) {
                 variance += ((data[row][col] - mean[col]) * (data[row][col] - mean[col]) - variance) / (row + 1);
             }
-            covarianceMatrix[index++] = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+            covarianceMatrix[index++] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
         }
 
         return covarianceMatrix;
     }
 
     @Override
-    public double[][] compute(boolean biasCorrected) {
-        double[][] covarianceMatrix = new double[numOfCols][numOfCols];
+    public float[][] compute(boolean biasCorrected) {
+        float[][] covarianceMatrix = new float[numOfCols][numOfCols];
 
-        double[] mean = computeMeans();
+        float[] mean = computeMeans();
 
         for (int col = 0; col < numOfCols; col++) {
             for (int col2 = 0; col2 < col; col2++) {
-                double variance = 0;
+                float variance = 0;
                 for (int row = 0; row < numOfRows; row++) {
                     variance += ((data[row][col] - mean[col]) * (data[row][col2] - mean[col2]) - variance) / (row + 1);
                 }
-                variance = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+                variance = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
                 covarianceMatrix[col][col2] = variance;
                 covarianceMatrix[col2][col] = variance;
             }
-            double variance = 0;
+            float variance = 0;
             for (int row = 0; row < numOfRows; row++) {
                 variance += ((data[row][col] - mean[col]) * (data[row][col] - mean[col]) - variance) / (row + 1);
             }
-            covarianceMatrix[col][col] = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+            covarianceMatrix[col][col] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
         }
 
         return covarianceMatrix;
