+++ tetrad-lib/src/main/java/edu/cmu/tetrad/stat/correlation/CovarianceMatrixForkJoin.java
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA 02110-1301  USA
  */
-package edu.cmu.tetrad.correlation;
+package edu.cmu.tetrad.stat.correlation;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -26,13 +26,13 @@ import java.util.concurrent.RecursiveAction;
 
 /**
  *
- * Jan 27, 2016 5:37:40 PM
+ * Jan 25, 2016 3:42:10 PM
  *
  * @author Kevin V. Bui (kvb2@pitt.edu)
  */
-public class RealCovarianceMatrixForkJoin implements RealCovariance {
+public class CovarianceMatrixForkJoin implements Covariance {
 
-    private final double[][] data;
+    private final float[][] data;
 
     private final int numOfRows;
 
@@ -42,7 +42,7 @@ public class RealCovarianceMatrixForkJoin implements RealCovariance {
 
     private final ForkJoinPool pool;
 
-    public RealCovarianceMatrixForkJoin(double[][] data, int numOfThreads) {
+    public CovarianceMatrixForkJoin(float[][] data, int numOfThreads) {
         this.data = data;
         this.numOfRows = data.length;
         this.numOfCols = data[0].length;
@@ -51,10 +51,10 @@ public class RealCovarianceMatrixForkJoin implements RealCovariance {
     }
 
     @Override
-    public double[] computeLowerTriangle(boolean biasCorrected) {
-        double[] covarianceMatrix = new double[(numOfCols * (numOfCols + 1)) / 2];
+    public float[] computeLowerTriangle(boolean biasCorrected) {
+        float[] covarianceMatrix = new float[(numOfCols * (numOfCols + 1)) / 2];
 
-        double[] means = new double[numOfCols];
+        float[] means = new float[numOfCols];
         pool.invoke(new MeanAction(means, data, 0, numOfCols - 1));
         pool.invoke(new CovarianceLowerTriangleAction(covarianceMatrix, means, 0, numOfCols - 1, biasCorrected));
 
@@ -62,10 +62,10 @@ public class RealCovarianceMatrixForkJoin implements RealCovariance {
     }
 
     @Override
-    public double[][] compute(boolean biasCorrected) {
-        double[][] covarianceMatrix = new double[numOfCols][numOfCols];
+    public float[][] compute(boolean biasCorrected) {
+        float[][] covarianceMatrix = new float[numOfCols][numOfCols];
 
-        double[] means = new double[numOfCols];
+        float[] means = new float[numOfCols];
         pool.invoke(new MeanAction(means, data, 0, numOfCols - 1));
         pool.invoke(new CovarianceAction(covarianceMatrix, means, 0, numOfCols - 1, biasCorrected));
 
@@ -76,13 +76,13 @@ public class RealCovarianceMatrixForkJoin implements RealCovariance {
 
         private static final long serialVersionUID = 1034920868427599720L;
 
-        private final double[][] covariance;
-        private final double[] means;
+        private final float[][] covariance;
+        private final float[] means;
         private final int start;
         private final int end;
         private final boolean biasCorrected;
 
-        public CovarianceAction(double[][] covariance, double[] means, int start, int end, boolean biasCorrected) {
+        public CovarianceAction(float[][] covariance, float[] means, int start, int end, boolean biasCorrected) {
             this.covariance = covariance;
             this.means = means;
             this.start = start;
@@ -93,19 +93,19 @@ public class RealCovarianceMatrixForkJoin implements RealCovariance {
         private void computeCovariance() {
             for (int col = start; col <= end; col++) {
                 for (int col2 = 0; col2 < col; col2++) {
-                    double variance = 0;
+                    float variance = 0;
                     for (int row = 0; row < numOfRows; row++) {
                         variance += ((data[row][col] - means[col]) * (data[row][col2] - means[col2]) - variance) / (row + 1);
                     }
-                    variance = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+                    variance = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
                     covariance[col][col2] = variance;
                     covariance[col2][col] = variance;
                 }
-                double variance = 0;
+                float variance = 0;
                 for (int row = 0; row < numOfRows; row++) {
                     variance += ((data[row][col] - means[col]) * (data[row][col] - means[col]) - variance) / (row + 1);
                 }
-                covariance[col][col] = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+                covariance[col][col] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
             }
         }
 
@@ -135,13 +135,13 @@ public class RealCovarianceMatrixForkJoin implements RealCovariance {
 
         private static final long serialVersionUID = 1818119309247848613L;
 
-        private final double[] covariance;
-        private final double[] means;
+        private final float[] covariance;
+        private final float[] means;
         private final int start;
         private final int end;
         private final boolean biasCorrected;
 
-        public CovarianceLowerTriangleAction(double[] covariance, double[] means, int start, int end, boolean biasCorrected) {
+        public CovarianceLowerTriangleAction(float[] covariance, float[] means, int start, int end, boolean biasCorrected) {
             this.covariance = covariance;
             this.means = means;
             this.start = start;
@@ -153,17 +153,17 @@ public class RealCovarianceMatrixForkJoin implements RealCovariance {
             int index = (start * (start + 1)) / 2;
             for (int col = start; col <= end; col++) {
                 for (int col2 = 0; col2 < col; col2++) {
-                    double variance = 0;
+                    float variance = 0;
                     for (int row = 0; row < numOfRows; row++) {
                         variance += ((data[row][col] - means[col]) * (data[row][col2] - means[col2]) - variance) / (row + 1);
                     }
-                    covariance[index++] = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+                    covariance[index++] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
                 }
-                double variance = 0;
+                float variance = 0;
                 for (int row = 0; row < numOfRows; row++) {
                     variance += ((data[row][col] - means[col]) * (data[row][col] - means[col]) - variance) / (row + 1);
                 }
-                covariance[index++] = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+                covariance[index++] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
             }
         }
 
@@ -193,12 +193,12 @@ public class RealCovarianceMatrixForkJoin implements RealCovariance {
 
         private static final long serialVersionUID = 2419217605658853345L;
 
-        private final double[] means;
-        private final double[][] data;
+        private final float[] means;
+        private final float[][] data;
         private final int start;
         private final int end;
 
-        public MeanAction(double[] means, double[][] data, int start, int end) {
+        public MeanAction(float[] means, float[][] data, int start, int end) {
             this.means = means;
             this.data = data;
             this.start = start;
@@ -207,7 +207,7 @@ public class RealCovarianceMatrixForkJoin implements RealCovariance {
 
         private void computeMean() {
             for (int col = start; col <= end; col++) {
-                double sum = 0;
+                float sum = 0;
                 for (int row = 0; row < numOfRows; row++) {
                     sum += data[row][col];
                 }
