+++ tetrad-lib/src/main/java/edu/cmu/tetrad/stat/correlation/CovarianceMatrixOnTheFly.java
@@ -16,15 +16,17 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA 02110-1301  USA
  */
-package edu.cmu.tetrad.correlation;
+package edu.cmu.tetrad.stat.correlation;
 
 /**
+ * Compute covariance on the fly. Warning! This class will overwrite the values
+ * in the input data.
  *
- * Jan 25, 2016 2:13:26 PM
+ * Jan 27, 2016 4:37:44 PM
  *
  * @author Kevin V. Bui (kvb2@pitt.edu)
  */
-public class CovarianceMatrix implements Covariance {
+public class CovarianceMatrixOnTheFly implements Covariance {
 
     private final float[][] data;
 
@@ -32,43 +34,30 @@ public class CovarianceMatrix implements Covariance {
 
     private final int numOfCols;
 
-    public CovarianceMatrix(float[][] data) {
+    public CovarianceMatrixOnTheFly(float[][] data) {
         this.data = data;
         this.numOfRows = data.length;
         this.numOfCols = data[0].length;
-    }
-
-    private float[] computeMeans() {
-        float[] mean = new float[numOfCols];
-        for (int col = 0; col < numOfCols; col++) {
-            float sum = 0;
-            for (int row = 0; row < numOfRows; row++) {
-                sum += data[row][col];
-            }
-            mean[col] = sum / numOfRows;
-        }
-
-        return mean;
     }
 
     @Override
     public float[] computeLowerTriangle(boolean biasCorrected) {
         float[] covarianceMatrix = new float[(numOfCols * (numOfCols + 1)) / 2];
 
-        float[] mean = computeMeans();
+        computeMeans();
 
         int index = 0;
         for (int col = 0; col < numOfCols; col++) {
             for (int col2 = 0; col2 < col; col2++) {
                 float variance = 0;
                 for (int row = 0; row < numOfRows; row++) {
-                    variance += ((data[row][col] - mean[col]) * (data[row][col2] - mean[col2]) - variance) / (row + 1);
+                    variance += ((data[row][col]) * (data[row][col2]) - variance) / (row + 1);
                 }
                 covarianceMatrix[index++] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
             }
             float variance = 0;
             for (int row = 0; row < numOfRows; row++) {
-                variance += ((data[row][col] - mean[col]) * (data[row][col] - mean[col]) - variance) / (row + 1);
+                variance += ((data[row][col]) * (data[row][col]) - variance) / (row + 1);
             }
             covarianceMatrix[index++] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
         }
@@ -80,13 +69,13 @@ public class CovarianceMatrix implements Covariance {
     public float[][] compute(boolean biasCorrected) {
         float[][] covarianceMatrix = new float[numOfCols][numOfCols];
 
-        float[] mean = computeMeans();
+        computeMeans();
 
         for (int col = 0; col < numOfCols; col++) {
             for (int col2 = 0; col2 < col; col2++) {
                 float variance = 0;
                 for (int row = 0; row < numOfRows; row++) {
-                    variance += ((data[row][col] - mean[col]) * (data[row][col2] - mean[col2]) - variance) / (row + 1);
+                    variance += ((data[row][col]) * (data[row][col2]) - variance) / (row + 1);
                 }
                 variance = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
                 covarianceMatrix[col][col2] = variance;
@@ -94,12 +83,25 @@ public class CovarianceMatrix implements Covariance {
             }
             float variance = 0;
             for (int row = 0; row < numOfRows; row++) {
-                variance += ((data[row][col] - mean[col]) * (data[row][col] - mean[col]) - variance) / (row + 1);
+                variance += ((data[row][col]) * (data[row][col]) - variance) / (row + 1);
             }
             covarianceMatrix[col][col] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
         }
 
         return covarianceMatrix;
+    }
+
+    private void computeMeans() {
+        for (int col = 0; col < numOfCols; col++) {
+            float mean = 0;
+            for (int row = 0; row < numOfRows; row++) {
+                mean += data[row][col];
+            }
+            mean /= numOfRows;
+            for (int row = 0; row < numOfRows; row++) {
+                data[row][col] -= mean;
+            }
+        }
     }
 
 }
