+++ tetrad-lib/src/main/java/edu/cmu/tetrad/stat/correlation/RealCovarianceMatrixForkJoinOnTheFly.java
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA 02110-1301  USA
  */
-package edu.cmu.tetrad.correlation;
+package edu.cmu.tetrad.stat.correlation;
 
 import java.util.LinkedList;
 import java.util.List;
