+++ tetrad-lib/src/main/java/edu/cmu/tetrad/stat/correlation/RealCovarianceMatrixOnTheFly.java
@@ -16,84 +16,31 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA 02110-1301  USA
  */
-package edu.cmu.tetrad.correlation;
+package edu.cmu.tetrad.stat.correlation;
 
 /**
- * Compute covariance on the fly. Warning! This class will overwrite the values
- * in the input data.
  *
- * Jan 27, 2016 4:37:44 PM
+ * Jan 27, 2016 5:48:23 PM
  *
  * @author Kevin V. Bui (kvb2@pitt.edu)
  */
-public class CovarianceMatrixOnTheFly implements Covariance {
+public class RealCovarianceMatrixOnTheFly implements RealCovariance {
 
-    private final float[][] data;
+    private final double[][] data;
 
     private final int numOfRows;
 
     private final int numOfCols;
 
-    public CovarianceMatrixOnTheFly(float[][] data) {
+    public RealCovarianceMatrixOnTheFly(double[][] data) {
         this.data = data;
         this.numOfRows = data.length;
         this.numOfCols = data[0].length;
     }
 
-    @Override
-    public float[] computeLowerTriangle(boolean biasCorrected) {
-        float[] covarianceMatrix = new float[(numOfCols * (numOfCols + 1)) / 2];
-
-        computeMeans();
-
-        int index = 0;
-        for (int col = 0; col < numOfCols; col++) {
-            for (int col2 = 0; col2 < col; col2++) {
-                float variance = 0;
-                for (int row = 0; row < numOfRows; row++) {
-                    variance += ((data[row][col]) * (data[row][col2]) - variance) / (row + 1);
-                }
-                covarianceMatrix[index++] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
-            }
-            float variance = 0;
-            for (int row = 0; row < numOfRows; row++) {
-                variance += ((data[row][col]) * (data[row][col]) - variance) / (row + 1);
-            }
-            covarianceMatrix[index++] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
-        }
-
-        return covarianceMatrix;
-    }
-
-    @Override
-    public float[][] compute(boolean biasCorrected) {
-        float[][] covarianceMatrix = new float[numOfCols][numOfCols];
-
-        computeMeans();
-
-        for (int col = 0; col < numOfCols; col++) {
-            for (int col2 = 0; col2 < col; col2++) {
-                float variance = 0;
-                for (int row = 0; row < numOfRows; row++) {
-                    variance += ((data[row][col]) * (data[row][col2]) - variance) / (row + 1);
-                }
-                variance = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
-                covarianceMatrix[col][col2] = variance;
-                covarianceMatrix[col2][col] = variance;
-            }
-            float variance = 0;
-            for (int row = 0; row < numOfRows; row++) {
-                variance += ((data[row][col]) * (data[row][col]) - variance) / (row + 1);
-            }
-            covarianceMatrix[col][col] = biasCorrected ? variance * ((float) numOfRows / (float) (numOfRows - 1)) : variance;
-        }
-
-        return covarianceMatrix;
-    }
-
     private void computeMeans() {
         for (int col = 0; col < numOfCols; col++) {
-            float mean = 0;
+            double mean = 0;
             for (int row = 0; row < numOfRows; row++) {
                 mean += data[row][col];
             }
@@ -102,6 +49,57 @@ public class CovarianceMatrixOnTheFly implements Covariance {
                 data[row][col] -= mean;
             }
         }
+    }
+
+    @Override
+    public double[] computeLowerTriangle(boolean biasCorrected) {
+        double[] covarianceMatrix = new double[(numOfCols * (numOfCols + 1)) / 2];
+
+        computeMeans();
+
+        int index = 0;
+        for (int col = 0; col < numOfCols; col++) {
+            for (int col2 = 0; col2 < col; col2++) {
+                double variance = 0;
+                for (int row = 0; row < numOfRows; row++) {
+                    variance += ((data[row][col]) * (data[row][col2]) - variance) / (row + 1);
+                }
+                covarianceMatrix[index++] = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+            }
+            double variance = 0;
+            for (int row = 0; row < numOfRows; row++) {
+                variance += ((data[row][col]) * (data[row][col]) - variance) / (row + 1);
+            }
+            covarianceMatrix[index++] = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+        }
+
+        return covarianceMatrix;
+    }
+
+    @Override
+    public double[][] compute(boolean biasCorrected) {
+        double[][] covarianceMatrix = new double[numOfCols][numOfCols];
+
+        computeMeans();
+
+        for (int col = 0; col < numOfCols; col++) {
+            for (int col2 = 0; col2 < col; col2++) {
+                double variance = 0;
+                for (int row = 0; row < numOfRows; row++) {
+                    variance += ((data[row][col]) * (data[row][col2]) - variance) / (row + 1);
+                }
+                variance = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+                covarianceMatrix[col][col2] = variance;
+                covarianceMatrix[col2][col] = variance;
+            }
+            double variance = 0;
+            for (int row = 0; row < numOfRows; row++) {
+                variance += ((data[row][col]) * (data[row][col]) - variance) / (row + 1);
+            }
+            covarianceMatrix[col][col] = biasCorrected ? variance * ((double) numOfRows / (double) (numOfRows - 1)) : variance;
+        }
+
+        return covarianceMatrix;
     }
 
 }
